azurite-init/requirements.txt
azure-storage-blob==12.19.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp==1.24.0
python-json-logger==2.0.7 